bak/server/ChickenDinner8/requirements.txt
django
uwsgi
mysqlclient
Pillow
pytz
python-weixin
simplejson